Refactored persistent data
Lots of method refactoring into CommandUtil or renaming. It appears to have not severely broken anything. Merging this is going to be a disaster, so commands should probably only be worked on in this branch until master is brought up to date
Tried to add some annotations and descriptions
Rephrased some CommandChannel return texts, added return for switching to the same channel